Add 13-binary_tree_balance

+++ 14-binary_tree_balance.c
@@ -9,8 +9,8 @@
 int binary_tree_balance(const binary_tree_t *tree)
 {
 	int l, r;
-	l = tree->left ? binary_tree_height(tree->left) : -1;
-	r = tree->right ? binary_tree_height(tree->right) : -1;
+	l = tree->left ? (int)binary_tree_height(tree->left) : -1;
+	r = tree->right ? (int)binary_tree_height(tree->right) : -1;
 	return (l - r);
 }
